Update README.md
The problem HEALTH TECH solves

Epidemological studies have become necessary for govts all around the world. But along with the vast data collected to analyse for such studies, a sense of privacy of citizens and anonymity is also necessary. This app helps to collect data from users like whether he/she infected or not. if he/she caused any diseases, they can use this app to update their current state to the medical departments through this app. Actually this app has only two buttons whether exposed (to disease) or not.
if it is exposed then a warning/alert in the form of red led is appeared in the users interface, if it is not then no alert is shown.
At the same time the data from every users is updated on thingSpeak.